V3

+++ Project2/Project2/OrderProcessor.cs
@@ -1,15 +1,17 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using System.Threading;
 
 namespace Project2
 {
-    public delegate void orderSuccess(string senderID, Int32 total, Int32 ticketPrice, Int32 numTickets);
+    public delegate void orderSuccess(string senderID, double total, double ticketPrice, int numTickets);
     public class OrderProcessor
     {
-         public static event orderSuccess OrderProcess;
+        public static event orderSuccess OrderProcess;
 
-        public static void orderProcessing(OrderClass orderObject, Int32 ticketPrice)
+        
+        public static void orderProcessing(OrderClass orderObject, double ticketPrice)
         {
 
             if(!checkCredit(orderObject.getcardNo()))
@@ -18,11 +20,11 @@ namespace Project2
             }
             else
             {
-                Int32 total = Convert.ToInt32(1.08 * (ticketPrice * orderObject.getAmount()));
+                double total = Convert.ToDouble(1.08 * (ticketPrice * orderObject.getAmount()));
                 OrderProcess(orderObject.getSenderID(), total, ticketPrice, orderObject.getAmount());
             }
         }
-
+        
         private static Boolean checkCredit(Int64 cardNum)
         {
             if (cardNum <= 10000000 && cardNum >= 99999999)
